requirements.txt
Flask
flask-cors
python-dotenv
openai
PyPDF2
python-docx
Werkzeug